+++ changelog.md
@@ -1,6 +1,7 @@
 # v4.20.0-dev
-- Added more normalization methods to `Hist1d`
+- Added more normalization methods to `Hist1d` and `Hist2d`
 - Added `bins` and `binwidth` properties to `Hist1d`
+- Added `xbins`, `ybins`, `xbinwidths`, `ybinwidths`, `binareas`, and `histogram` properties to `Hist2d`
 
 # v4.19.0
 - Added ability to have `*` in `savefig`